tax_automator/requirements.txt
google-adk
firebase-admin
google-cloud-firestore
fastapi
uvicorn
